1. getExtension支持根据传入构造方法参数来匹配构造方法 2. 修复code扩容遗留bug

+++ src/test/java/org/kin/framework/utils/ExtensionLoaderTest.java
@@ -7,7 +7,7 @@ package org.kin.framework.utils;
 public class ExtensionLoaderTest {
     public static void main(String[] args) {
         System.out.println(ExtensionLoader.getExtensions(KinService.class));
-        System.out.println(ExtensionLoader.getExtension(KinService.class, (byte) 1));
+        System.out.println(ExtensionLoader.getExtension(KinService.class, 1));
         System.out.println(ExtensionLoader.getExtension(KinService.class, "E"));
         System.out.println(ExtensionLoader.getExtensionOrDefault(KinService.class, "E"));
